chore: add commitment.go with temporary commit acknowledgement func
